Adding in inverse rotation and rotation for sim code.

+++ apps/hmc5883/fsw/src/hmc5883_custom.h
@@ -165,7 +165,7 @@ boolean HMC5883_Custom_Measure(int16 *X, int16 *Y, int16 *Z);
 **  \returns    TRUE for success, FALSE for failure.
 **
 *************************************************************************/
-boolean HMC5883_Apply_Platform_Rotation(int16 *X, int16 *Y, int16 *Z);
+boolean HMC5883_Apply_Platform_Rotation(float *X, float *Y, float *Z);
 
 
 /************************************************************************/
